// src/api/generated/models/versionDataDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Flowforge API
 * The Flowforge workflow orchestration engine API documentation
 * OpenAPI spec version: 1.0
 */
import type { VersionCreatorDto } from './versionCreatorDto';
import type { VersionDataDtoDefinition } from './versionDataDtoDefinition';

export interface VersionDataDto {
  id: string;
  workflowId: string;
  version: number;
  definition: VersionDataDtoDefinition;
  isActive: boolean;
  /** Topological execution order (returned on version creation) */
  executionOrder?: string[];
  creator: VersionCreatorDto;
  createdAt: string;
}